fix(booking): adjust status verification to reopen room calendar upon booking cancellation

+++ src/main/java/com/stayhub/backend/Module/Property/DTO/Response/MySubscriptionResponse.java
@@ -12,7 +12,7 @@ public record MySubscriptionResponse(
         SubscriptionTier tier,
         String planName,
         Integer maxListings,
-        Double commissionRate,
+        BigDecimal commissionRate,
         BigDecimal creditLimit,
         LocalDateTime startDate,
         LocalDateTime endDate,
